feat(explorer): use client wasm as late as possible
So they are never included in Server side renders.
